+++ src/components/index.js
@@ -1,6 +1,6 @@
 export { default as HomePage } from './HomePage';
 export { default as AccountPage } from './AccountPage';
-export { default as SignupPage } from './SignupPage';
-export { default as SigninPage } from './SigninPage';
+export { default as SignupPage } from './Signup';
+export { default as SigninPage } from './Login';
 export { default as PasswordForgetPage } from './PasswordForgetPage';
 export { default as Landing } from './Landing';
